+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.1]
+
+* fix search field name
+* fix reloading of map after saving
+* code cleanup
+
 ## [3.0.0]
 
 * Upgrade for Silverstripe 6 (thanks @lerni)